+++ src/hooks/query/useGetCommentList.ts
@@ -1,5 +1,5 @@
 import getCommentList from '@/api/findGuideDetail/getCommentList';
-import { COMMENT_LIST } from '@/constants/queyKey';
+import { COMMENT_LIST } from '@/constants/queryKey';
 import { useQuery } from '@tanstack/react-query';
 
 const useGetCommentList = (guidePostId: string, page: number) => {
