+++ src/hooks/query/useGetMyJoinedTravel.ts
@@ -1,5 +1,5 @@
 import getMyJoinedTravel from '@/api/myJoinedTravel/getMyJoinedTravel';
-import { MY_JOINED_TRAVEL } from '@/constants/queyKey';
+import { MY_JOINED_TRAVEL } from '@/constants/queryKey';
 import { useQuery } from '@tanstack/react-query';
 
 const useGetMyJoinedTravel = (userId: string) => {
